Tiny memory optimization: u8 for age_years, not u64. #1
Memory usage didn't actually budge -- still 13.52GiB. But this should
certainly help serialized size.

+++ src/make_population.rs
@@ -156,7 +156,7 @@ struct TuPerson {
     pleisure: f64,
     pshop: f64,
     pschool: f64,
-    age: usize,
+    age: u8,
 }
 
 fn parse_usize_or_na<'de, D: Deserializer<'de>>(d: D) -> Result<Option<usize>, D::Error> {
